initial logic move
